+++ Tests/WindowsCloudPC.Tests.ps1
@@ -2,6 +2,7 @@ BeforeDiscovery {
     $script:PublicFunctions = @(
         'Connect-CloudPC',
         'Get-CloudPC',
+        'Get-CloudPCByProvisioningPolicy',
         'Get-CloudPCProvisioningPolicy',
         'Get-CloudPCUsage'
     )